pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal_vein_fusion"
version = "0.1.0"
description = "Multimodal (RGB/NIR/Thermal) vein fusion with Tkinter GUI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.23",
  "opencv-python>=4.8",
  "pillow>=10.0",
]

[project.optional-dependencies]
skimage = ["scikit-image>=0.21"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["src"]
